==> baseline/lassosum2/1_input_data.ipynb <==
# get the EAS combine pheno and covar
import pandas as pd
import os
import numpy as np

# /data1/jiapl_group/lishuhua/project/PRS_benchmark/real_data/Cross_Validation/CAS/alt/group_1/ids/combined_ids.txt
eas_base_dir = "/data1/jiapl_group/lishuhua/project/PRS_benchmark/real_data/Cross_Validation/CAS/"
eas_covar_path = "/data1/jiapl_group/lishuhua/project/PRS_benchmark/real_data/CAS/pheno/covariates.txt"
output_base_dir = "/data1/jiapl_group/lishuhua/project/PRS_benchmark/software/lassosum/test/EAS/"

trait_list = ['waist', 'height', 'pulse', 'dbp', 'sbp', 'smoke', 'drink', 'bmi', 'wbc', 'rbc', 'hb', 'plt', 'lymph', 'mono', 'neut', 'eos', 'alt', 'ast', 'bun', 'cholesterol', 'creatinine', 'glucose', 'ggt', 'hdl', 'ldl', 'triglycerides', 'ua']

for trait in trait_list:
    for group in range(1, 11):
        print(f"Processing {trait} group {group}")
        ids_path = os.path.join(eas_base_dir, trait, f"group_{group}", "ids", "combined_ids.txt")
        pheno_path_1 = os.path.join(eas_base_dir, trait, f"group_{group}", "pheno", "test_pheno.txt")
        pheno_path_2 = os.path.join(eas_base_dir, trait, f"group_{group}", "pheno", "tune_pheno.txt")
        ids_df = pd.read_csv(ids_path, header=None, names=["FID", "IID"], sep="\t")
        pheno_df_1 = pd.read_csv(pheno_path_1, sep="\t")
        pheno_df_2 = pd.read_csv(pheno_path_2, sep="\t")
        pheno_df = pd.concat([pheno_df_1, pheno_df_2], ignore_index=True)
        print(f"IDs shape: {ids_df.shape}, Pheno shape: {pheno_df.shape}")
        covar_df = pd.read_csv(eas_covar_path, sep="\t")
        ids_df["FID"] = ids_df["FID"].astype(str)
        ids_df["IID"] = ids_df["IID"].astype(str)
        pheno_df["FID"] = pheno_df["FID"].astype(str)
        pheno_df["IID"] = pheno_df["IID"].astype(str)
        covar_df["FID"] = covar_df["FID"].astype(str)
        covar_df["IID"] = covar_df["IID"].astype(str)
        merged_df = ids_df.merge(pheno_df, on=["FID", "IID"]).merge(covar_df, on=["FID", "IID"])
        merged_pheno_df = ids_df.merge(pheno_df, on=["FID", "IID"])
        merged_covar_df = ids_df.merge(covar_df, on=["FID", "IID"])
        if merged_pheno_df.shape[0] != ids_df.shape[0]:
            print(f"Warning: Mismatch in number of IDs for {trait} group {group} in pheno")
        if merged_covar_df.shape[0] != ids_df.shape[0]:
            print(f"The shape of covar: {merged_covar_df.shape}, IDs shape: {ids_df.shape}")
            print(f"Warning: Mismatch in number of IDs for {trait} group {group} in covar")
        if merged_df.shape[0] != ids_df.shape[0]:
            print(f"Warning: Mismatch in number of IDs for {trait} group {group}")
        output_path = os.path.join(output_base_dir, trait, f"group_{group}")
        os.makedirs(output_path, exist_ok=True)
        merged_df.to_csv(os.path.join(output_path, "pheno_covar.txt"), sep="\t", index=False)
        merged_pheno_df.to_csv(os.path.join(output_path, "pheno.txt"), sep="\t", index=False)
        merged_covar_df.to_csv(os.path.join(output_path, "covar.txt"), sep="\t", index=False)

# get the EUR combine pheno and covar
import pandas as pd
import os
import numpy as np

# /data1/jiapl_group/lishuhua/project/PRS_benchmark/real_data/Cross_Validation/UKB_EUR/pheno/fold_1/tune_ids.txt
eur_base_dir = "/data1/jiapl_group/lishuhua/project/PRS_benchmark/real_data/Cross_Validation/UKB_EUR/"
output_base_dir = "/data1/jiapl_group/lishuhua/project/PRS_benchmark/software/lassosum/test/EUR/"
covar_list = ["age", "sex", "PC1", "PC2", "PC3", "PC4", "PC5", "PC6", "PC7", "PC8", "PC9", "PC10", "PC11", "PC12", "PC13", "PC14", "PC15", "PC16", "PC17", "PC18", "PC19", "PC20"]

trait_dict = {
        'p48': 'waist',
        'p50': 'height',
        'p102': 'pulse',
        'p4079': 'dbp',
        'p4080': 'sbp',
        'p20116': 'smoke',
        'p20117': 'drink',
        'p21001': 'bmi',
        'p30000': 'wbc',
        'p30010': 'rbc',
        'p30020':'hb',
        'p30080': 'plt',
        'p30120': 'lymph',
        'p30130': 'mono',
        'p30140': 'neut',
        'p30150': 'eos',
        'p30620': 'alt',
        'p30650': 'ast',
        'p30670': 'bun',
        'p30690': 'cholesterol',
        'p30700': 'creatinine',
        'p30730': 'ggt',
        'p30740': 'glucose',
        'p30760': 'hdl',
        'p30780': 'ldl',
        'p30870': 'triglycerides',
        'p30880': 'ua'
    }

for group in range(1, 11):
    ids_path = os.path.join(eur_base_dir, f"pheno/fold_{group}", "tune_ids.txt")
    covars_path = os.path.join(eur_base_dir, f"pheno/fold_{group}", "tune_covar.txt")
    pheno_path = os.path.join(eur_base_dir, f"pheno/fold_{group}", "tune_pheno.txt")
    ids_df = pd.read_csv(ids_path, header=None, names=["FID", "IID"], sep="\t")
    pheno_df = pd.read_csv(pheno_path, sep="\t")
    covar_df = pd.read_csv(covars_path, sep="\t")
    covar_df = covar_df[["FID", "IID"] + covar_list]
    print(f"IDs shape: {ids_df.shape}, Pheno shape: {pheno_df.shape}, Covar shape: {covar_df.shape}")
    ids_df["FID"] = ids_df["FID"].astype(str)
    ids_df["IID"] = ids_df["IID"].astype(str)
    covar_df["FID"] = covar_df["FID"].astype(str)
    covar_df["IID"] = covar_df["IID"].astype(str)
    pheno_df["FID"] = pheno_df["FID"].astype(str)
    pheno_df["IID"] = pheno_df["IID"].astype(str)
    merge_df = ids_df.merge(pheno_df, on=["FID", "IID"]).merge(covar_df, on=["FID", "IID"])
    if merge_df.shape[0] != ids_df.shape[0]:
        print(f"Warning: Mismatch in number of IDs for group {group}")
    for trait_id, trait in trait_dict.items():
        print(f"Processing {trait} group {group}")
        pheno_name = f"{trait_id}_raw"
        if trait in ['smoke', 'drink']:
            pheno_name = f"{trait_id}_int"
        merged_df = merge_df[["FID", "IID", pheno_name] + [col for col in covar_df.columns if col not in ["FID", "IID"]]]
        # rename the column pheno_name to "Pheno"
        merged_df.rename(columns={pheno_name: "Pheno"}, inplace=True)
        output_path = os.path.join(output_base_dir, trait, f"group_{group}")
        os.makedirs(output_path, exist_ok=True)
        merged_df.to_csv(os.path.join(output_path, "pheno_covar.txt"), sep="\t", index=False)
        merged_df[["FID", "IID", "Pheno"]].to_csv(os.path.join(output_path, "pheno.txt"), sep="\t", index=False)
        merged_df[["FID", "IID"] + [col for col in covar_df.columns if col not in ["FID", "IID"]]].to_csv(os.path.join(output_path, "covar.txt"), sep="\t", index=False)
